+++ src/App.js
@@ -9,7 +9,7 @@ import { withAuthentication } from './components/Session'
 import LogIn from "./components/Application/LogIn";
 import SignUp from "./components/Application/SignUp";
 import LogOut from "./components/Application/LogOut";
-import AppPulpit from "./components/Application/AppPulpit";
+import AppHome from "./components/Application/AppHome";
 
 
 function App() {
@@ -21,7 +21,7 @@ function App() {
           <Route path={'/login'} exact component={LogIn}/>
           <Route path={'/signup'} exact component={SignUp}/>
           <Route path={'/logout'} exact component={LogOut}/>
-          <Route path={'/app'} exact component={AppPulpit}/>
+          <Route path={'/app'} exact component={AppHome}/>
         </Switch>
       </Router>
   );
